Add a `timestamp()` method (and verification)
The `Client` object now exposes both a timestamp creation and
verification routine. The timestamp routine creates a Unix timestamp and
signs it with the client's private key.

The verification routine splits the generated timestamp from its linked
signature and verifies the signature against a specified public key.

These changes are made to support adding signatures to HTTP request
headers by generating a signed timestamp and verifying it with another
SDK. Other headers can be arbitrarily signed as well as the
functionality now exposes signatures and verification for arbitrary
strings.